Module bigtwo::trick

Represents the current state of a Trick, with logic to step through the trick, collecting player inputs, and progressing the Trick. NOTE: A single Game is composed of a series of Tricks.

The Caller is expected to keep track of the Players, provide the Players with Cards, and keep track of which player is supposed to begin the Trick.

Use with is fn perform_trick(…) -> TrickResult

Modules

check_player_can_play_hand 🔒

Checks if a specified Player can actually play the Hand they are attempting to play.

next_player_id 🔒

Identifies the Player who is up next.

Structs

Trick 🔒

Represents the current state of a Trick, keeps track of which hands have been played and who has passed, and who is the current player.

Enums

StepStatus 🔒

Returned at the end of each Player’s turn, informs the caller whether the Trick has ended (and how), or ig the Trick continues

TrickResult

Returned at the end of a Trick to signify to the caller

Functions

perform_trick

Performs the entire Trick and returns the TrickResult. It is expected that the caller will keep calling this until it produces a TrickResult::GameOver.