Function bigtwo::trick::next_player_id::next_player_id

pub fn next_player_id(
    current_player_id: usize,
    passed_player_ids: &BTreeSet<usize>
) -> usize

Identifies the Player who is up next.

Panics:

Panics if current_player_id < NUM_PLAYERS, since the id is actually an idx into [Player; 4].

Panics if there are 0 players who have not passed, that is not a valid game state (Trick ends when only 1 remaining player has not passed).

Panics if the computed next_id is equal to the current_player_id